First commit, not finished. Message and Database objects working well.
M is a message object, which includes handy message creation, in-place
creation of a proof-of-work token, and string representation or list
export

D is a database object, which includes import/export of databases from
json serialised files, message import with proof-of-work verification,
regex searching of messages, message lookup by hash and message
presentation as a date/time sorted list, ready for printing.

Necessary work:
1) D object needs code to add/remove "follows" (subscribed regexes)
and "peers" (other servers to poll and inform connecting peers of).
2) Need to write XML server function, hopefully a brainfuck oneliner
like that found in TinyP2P.
3) Need to write the actual script body that executes server or client
functions based on command line arguments.